# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Simplon Sénégal",
        "description": "Cet API permet de gérer les candidatures que Simplon Sénégal reçoit pour leurs différentes formations",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/candidatures": {
            "get": {
                "tags": [
                    "Candidatures"
                ],
                "summary": "Lister toutes les candidatures",
                "operationId": "250a566b2d074990f65e363d8b4b3d2c",
                "responses": {
                    "200": {
                        "description": "Succès",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": ""
                            }
                        }
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
        "/api/users/candidatures": {
            "get": {
                "tags": [
                    "Candidatures"
                ],
                "summary": "Lister les candidatures de l'utilisateur connecté",
                "operationId": "222ac7fe6f4cb188d7a56f3dbbd3c724",
                "responses": {
                    "200": {
                        "description": "Succès",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": ""
                            }
                        }
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
        "/api/acceptees": {
            "get": {
                "tags": [
                    "Candidatures"
                ],
                "summary": "Lister les candidatures acceptées",
                "operationId": "15b1b808816f6fc9af417e0485b7ceec",
                "responses": {
                    "200": {
                        "description": "Succès",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": ""
                            }
                        }
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
        "/api/en_attente": {
            "get": {
                "tags": [
                    "Candidatures"
                ],
                "summary": "Lister les candidatures en attente",
                "operationId": "0d5f7f7e5aa92cc4e589b88f29fefa0d",
                "responses": {
                    "200": {
                        "description": "Succès",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": ""
                            }
                        }
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
        "/api/refusees": {
            "get": {
                "tags": [
                    "Candidatures"
                ],
                "summary": "Lister les candidatures refusées",
                "operationId": "7af4ac1292a16b28fd02385a483f8ac6",
                "responses": {
                    "200": {
                        "description": "Succès",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": ""
                            }
                        }
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
        "/api/formations/candidatures/{formation}": {
            "get": {
                "tags": [
                    "Candidatures"
                ],
                "summary": "Ajouter une candidature",
                "operationId": "b4a1cef4e7ed3e20ea3fbd7ce172f33f",
                "parameters": [
                    {
                        "name": "formation",
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Succès",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": ""
                            }
                        }
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
        "/api/candidature/{candidature}": {
            "get": {
                "tags": [
                    "Candidatures"
                ],
                "summary": "Obtenir les détails d'une candidature",
                "operationId": "8ab217bc84eeb33f6e1d68c48b26c63e",
                "parameters": [
                    {
                        "name": "candidature",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Succès",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": ""
                            }
                        }
                    },
                    "404": {
                        "description": "Candidature non trouvée",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": ""
                            }
                        }
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
        "/api/candidatures/refuser/{candidature}": {
            "put": {
                "tags": [
                    "Candidatures"
                ],
                "summary": "Refuser une candidature",
                "operationId": "26d358197710c8142b28d0d8c78c138f",
                "parameters": [
                    {
                        "name": "candidature",
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Succès",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": ""
                            }
                        }
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
        "/api/candidatures/accepter/{candidature}": {
            "put": {
                "tags": [
                    "Candidatures"
                ],
                "summary": "Accepter une candidature",
                "operationId": "53d8a26de5f5b6a6d4667ab9fe2d6eab",
                "parameters": [
                    {
                        "name": "candidature",
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Succès",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": ""
                            }
                        }
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "BearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}